+++ main/src/main/java/net/oneandone/stool/server/configuration/StageConfiguration.java
@@ -49,10 +49,6 @@ public class StageConfiguration {
     @Option(key = "quota")
     public int quota;
 
-    /** max ram for container, in MB */
-    @Option(key = "memory")
-    public Integer memory;
-
     /** login names or email addresses, or "@last-modified-by" or "@created-by" */
     @Option(key = "notify")
     public List<String> notify;
@@ -71,7 +67,6 @@ public class StageConfiguration {
         this.notify = new ArrayList<>();
         this.notify.add(NOTIFY_CREATED_BY);
         this.quota = 10000;
-        this.memory = 400;
         this.expire = Expire.never();
         this.url = "(http:https)://%h/";
         this.comment = "";