+++ main/src/main/java/net/oneandone/stool/server/util/Session.java
@@ -269,23 +269,15 @@ public class Session {
     private int memReservedContainers() throws IOException {
         int reserved;
         Engine engine;
-        StageConfiguration stage;
         JsonObject json;
         Image image;
-        FileNode directory;
 
         reserved = 0;
         engine = dockerEngine();
         for (String container : engine.containerListRunning(Stage.CONTAINER_LABEL_STOOL).keySet()) {
             json = engine.containerInspect(container, false);
             image = Image.load(engine, Strings.removeLeft(json.get("Image").getAsString(), "sha256:"));
-            directory = stages.join(image.id);
-            if (directory.exists()) {
-                stage = loadStageConfiguration(directory);
-                reserved += stage.memory;
-            } else {
-                // TODO: stage container?
-            }
+            reserved += image.memory;
         }
         return reserved;
     }